+++ database/migrations/0001_01_01_000002_create_jobs_table.php
@@ -4,8 +4,7 @@ use Illuminate\Database\Migrations\Migration;
 use Illuminate\Database\Schema\Blueprint;
 use Illuminate\Support\Facades\Schema;
 
-return new class extends Migration
-{
+return new class extends Migration {
     /**
      * Run the migrations.
      */
@@ -50,8 +49,10 @@ return new class extends Migration
      */
     public function down(): void
     {
-        Schema::dropIfExists('jobs');
-        Schema::dropIfExists('job_batches');
-        Schema::dropIfExists('failed_jobs');
+        if (app()->isLocal()) {
+            Schema::dropIfExists('jobs');
+            Schema::dropIfExists('job_batches');
+            Schema::dropIfExists('failed_jobs');
+        }
     }
 };
