+++ database/migrations/2025_09_22_085310_create_telescope_entries_table.php
@@ -4,8 +4,7 @@ use Illuminate\Database\Migrations\Migration;
 use Illuminate\Database\Schema\Blueprint;
 use Illuminate\Support\Facades\Schema;
 
-return new class extends Migration
-{
+return new class extends Migration {
     /**
      * Get the migration connection name.
      */
@@ -61,10 +60,12 @@ return new class extends Migration
      */
     public function down(): void
     {
-        $schema = Schema::connection($this->getConnection());
+        if (app()->isLocal()) {
+            $schema = Schema::connection($this->getConnection());
 
-        $schema->dropIfExists('telescope_entries_tags');
-        $schema->dropIfExists('telescope_entries');
-        $schema->dropIfExists('telescope_monitoring');
+            $schema->dropIfExists('telescope_entries_tags');
+            $schema->dropIfExists('telescope_entries');
+            $schema->dropIfExists('telescope_monitoring');
+        }
     }
 };
